backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
gunicorn==23.0.0
whitenoise==6.9.0
psycopg[binary]==3.2.7
dj-database-url==2.3.0
python-dotenv==1.0.1
Pillow==11.2.1
